refactor(core): separate containers into own modules
