+++ packages/config-eslint/index.js
@@ -2,7 +2,7 @@ module.exports = {
     extends: ['alloy', 'alloy/typescript'],
     overrides: [
         {
-            files: ['**/*.spec.ts'],
+            files: ['**/*.spec.ts', '**/__tests__/*.ts'],
             env: {
                 jest: true,
             },
