+++ packages/core/src/containers/enum.ts
@@ -1,0 +1,138 @@
+import { concatUint8Arrays } from '@scale-codec/util';
+import { Codec, DecodeResult } from '../types';
+
+export class RustEnum<V> {
+    public readonly variant: keyof V;
+
+    public readonly value: unknown;
+
+    public constructor(variantName: keyof V, value?: unknown) {
+        this.value = value;
+        this.variant = variantName;
+    }
+
+    public is<K extends keyof V>(variant: K): boolean {
+        return this.variant === variant;
+    }
+
+    public as<K extends { [x in keyof V]: V[x] extends null ? never : x }[keyof V]>(variant: K): V[K] {
+        if (this.is(variant)) {
+            return this.value as V[K];
+        }
+
+        throw new Error(`cast failed - enum is not the "${variant}"`);
+    }
+
+    public match<R = any>(matchMap: RustEnumMatchMap<V, R>): R {
+        return matchMap[this.variant](this.value as any);
+    }
+    public toJSON() {
+        return {
+            variant: this.variant,
+            value: this.value,
+        };
+    }
+}
+
+export type RustEnumMatchMap<V, R = any> = {
+    [K in keyof V]: V[K] extends null ? () => R : (value: V[K]) => R;
+};
+
+export type RustEnumSchemaDef<V> = {
+    [K in keyof V]: { discriminant: number };
+};
+
+export class RustEnumSchema<V> {
+    private readonly def: RustEnumSchemaDef<V>;
+
+    private readonly disVarMap: Record<number, keyof V>;
+
+    public constructor(def: RustEnumSchemaDef<V>) {
+        this.def = def;
+        this.disVarMap = Object.fromEntries(
+            (Object.entries(def) as [keyof V, { discriminant: number }][]).map(([variant, { discriminant }]) => [
+                discriminant,
+                variant,
+            ]),
+        );
+
+        this.getVariantDiscriminant = this.getVariantDiscriminant.bind(this);
+        this.getDiscriminantVariant = this.getDiscriminantVariant.bind(this);
+        this.enumCodec = this.enumCodec.bind(this);
+        this.create = this.create.bind(this);
+    }
+
+    public getVariantDiscriminant(variant: keyof V): number {
+        return this.def[variant].discriminant;
+    }
+
+    public getDiscriminantVariant(discriminant: number): keyof V {
+        return this.disVarMap[discriminant];
+    }
+
+    public enumCodec(codecs: RustEnumCodecs<V>): RustEnumCodec<V> {
+        return new RustEnumCodec(this, codecs);
+    }
+
+    public create<K extends { [x in keyof V]: V[x] extends null ? x : never }[keyof V]>(emptyVariant: K): RustEnum<V>;
+    public create<K extends { [x in keyof V]: V[x] extends null ? never : x }[keyof V]>(
+        variantWithValue: K,
+        // eslint-disable-next-line @typescript-eslint/unified-signatures
+        value: V[K],
+    ): RustEnum<V>;
+    public create(variant: keyof V, value?: any): RustEnum<V> {
+        return new RustEnum(variant, value ?? null);
+    }
+}
+
+export type RustEnumNonEmptyVariants<V> = {
+    [K in keyof V]: V[K] extends null ? never : K;
+}[keyof V];
+
+export type RustEnumCodecs<V> = {
+    [K in RustEnumNonEmptyVariants<V>]: Codec<V[K]>;
+};
+
+export class RustEnumCodec<V> implements Codec<RustEnum<V>> {
+    private schema: RustEnumSchema<V>;
+    private codecs: RustEnumCodecs<V>;
+
+    public constructor(schema: RustEnumSchema<V>, codecs: RustEnumCodecs<V>) {
+        this.schema = schema;
+        this.codecs = codecs;
+        this.encode = this.encode.bind(this);
+        this.decode = this.decode.bind(this);
+    }
+
+    public encode(val: RustEnum<V>): Uint8Array {
+        const { variant, value } = val;
+        const discriminant = this.schema.getVariantDiscriminant(variant);
+        const codec = this.codecByVariant(variant);
+
+        const arrs: Uint8Array[] = [new Uint8Array([discriminant])];
+        codec && arrs.push(codec.encode(value));
+
+        return concatUint8Arrays(arrs);
+    }
+
+    public decode(bytes: Uint8Array): DecodeResult<RustEnum<V>> {
+        const discriminant = bytes[0];
+        const variant = this.schema.getDiscriminantVariant(discriminant);
+        const codec = this.codecByVariant(variant);
+
+        let len = 1;
+        let value: unknown = null;
+
+        if (codec) {
+            const [decoded, decodedLen] = codec.decode(bytes.subarray(1));
+            len += decodedLen;
+            value = decoded;
+        }
+
+        return [new RustEnum(variant, value), len];
+    }
+
+    private codecByVariant(variant: keyof V): Codec<unknown> | null {
+        return variant in this.codecs ? this.codecs[variant as keyof RustEnumCodecs<V>] : null;
+    }
+}
